Change test for plot_coord
